requirements.txt
streamlit
pandas
python-dotenv
google-api-python-client
google-generativeai
